+++ src/main/java/matal/config/JpaConfig.java
@@ -4,7 +4,7 @@ import org.springframework.context.annotation.Configuration;
 import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
 import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
 
-@EnableJpaRepositories
+@EnableJpaRepositories(basePackages = {"matal"})
 @EnableJpaAuditing
 @Configuration
 public class JpaConfig {
