+++ src/main/java/matal/member/entity/Member.java
@@ -1,4 +1,4 @@
-package matal.entity;
+package matal.member.entity;
 
 import jakarta.persistence.*;
 import lombok.Builder;
