+++ src/main/java/matal/store/repository/ReviewRepository.java
@@ -1,8 +1,10 @@
-package matal.repository;
+package matal.store.repository;
 
-import matal.entity.Review;
+import matal.store.entity.Review;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
 
+@Repository
 public interface ReviewRepository extends JpaRepository<Review, Long> {
     
 }
